+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_unnecessary_setstate/examples/example.dart
@@ -37,6 +37,17 @@ class _MyWidgetState extends State<MyWidget> {
     });
   }
 
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+
+    // LINT
+    setState(() {
+      myString = "Hello";
+    });
+    myMethod(); // LINT
+  }
+
   void myMethod() {
     setState(() {
       myString = "Hello";