+++ test/src/analyzers/lint_analyzer/rules/rules_list/no_equal_arguments/no_equal_arguments_rule_test.dart
@@ -9,6 +9,8 @@ const _incorrectExamplePath =
     'no_equal_arguments/examples/incorrect_example.dart';
 const _namedParametersExamplePath =
     'no_equal_arguments/examples/named_parameters_example.dart';
+const _providerExamplePath =
+    'no_equal_arguments/examples/provider_example.dart';
 
 void main() {
   group('NoEqualArgumentsRule', () {
@@ -58,33 +60,69 @@ void main() {
 
       RuleTestHelper.verifyNoIssues(issues);
     });
+
+    test('reports about found issue with default config', () async {
+      final unit =
+          await RuleTestHelper.resolveFromFile(_namedParametersExamplePath);
+      final issues = NoEqualArgumentsRule().check(unit);
+
+      RuleTestHelper.verifyIssues(
+        issues: issues,
+        startLines: [9],
+        startColumns: [3],
+        locationTexts: ['lastName: firstName'],
+        messages: ['The argument has already been passed.'],
+      );
+    });
+
+    test('reports no issues for provider read', () async {
+      final unit = await RuleTestHelper.resolveFromFile(_providerExamplePath);
+      final issues = NoEqualArgumentsRule().check(unit);
+
+      RuleTestHelper.verifyNoIssues(issues);
+    });
+
+    test('reports no issues with custom config', () async {
+      final unit =
+          await RuleTestHelper.resolveFromFile(_namedParametersExamplePath);
+      final config = {
+        'ignored-parameters': [
+          'lastName',
+        ],
+      };
+
+      final issues = NoEqualArgumentsRule(config).check(unit);
+
+      RuleTestHelper.verifyNoIssues(issues);
+    });
   });
 
-  test('reports about found issue with default config', () async {
-    final unit =
-        await RuleTestHelper.resolveFromFile(_namedParametersExamplePath);
-    final issues = NoEqualArgumentsRule().check(unit);
+  test('reports about found issue with ignored arguments config', () async {
+    final unit = await RuleTestHelper.resolveFromFile(_incorrectExamplePath);
+    final issues = NoEqualArgumentsRule({
+      'ignored-arguments': ['firstName'],
+    }).check(unit);
 
     RuleTestHelper.verifyIssues(
       issues: issues,
-      startLines: [9],
-      startColumns: [3],
-      locationTexts: ['lastName: firstName'],
-      messages: ['The argument has already been passed.'],
-    );
-  });
-
-  test('reports no issues with custom config', () async {
-    final unit =
-        await RuleTestHelper.resolveFromFile(_namedParametersExamplePath);
-    final config = {
-      'ignored-parameters': [
-        'lastName',
+      startLines: [32, 37, 42, 47, 54, 59],
+      startColumns: [5, 5, 5, 5, 5, 5],
+      locationTexts: [
+        'user.firstName',
+        'user.getName',
+        'user.getFirstName()',
+        "user.getNewName('name')",
+        'user.getNewName(name)',
+        'lastName: user.firstName',
       ],
-    };
-
-    final issues = NoEqualArgumentsRule(config).check(unit);
-
-    RuleTestHelper.verifyNoIssues(issues);
+      messages: [
+        'The argument has already been passed.',
+        'The argument has already been passed.',
+        'The argument has already been passed.',
+        'The argument has already been passed.',
+        'The argument has already been passed.',
+        'The argument has already been passed.',
+      ],
+    );
   });
 }